+++ src/main/java/binnie/craftgui/controls/page/ControlPage.java
@@ -8,10 +8,10 @@ public class ControlPage<T> extends Control implements IControlValue<T> {
 	T value;
 
 	public ControlPage(final IWidget parent, final T value) {
-		this(parent, 0.0f, 0.0f, parent.w(), parent.h(), value);
+		this(parent, 0, 0, parent.w(), parent.h(), value);
 	}
 
-	public ControlPage(final IWidget parent, final float x, final float y, final float w, final float h, final T value) {
+	public ControlPage(final IWidget parent, final int x, final int y, final int w, final int h, final T value) {
 		super(parent, x, y, w, h);
 		this.setValue(value);
 		if (parent instanceof IControlValue && ((IControlValue) parent).getValue() == null) {
