+++ src/main/java/binnie/craftgui/controls/scroll/ControlScroll.java
@@ -6,7 +6,7 @@ import binnie.craftgui.core.IWidget;
 public class ControlScroll extends Control {
 	private IControlScrollable scrollWidget;
 
-	public ControlScroll(final IWidget parent, final float x, final float y, final float width, final float height, final IControlScrollable scrollWidget) {
+	public ControlScroll(final IWidget parent, final int x, final int y, final int width, final int height, final IControlScrollable scrollWidget) {
 		super(parent, x, y, width, height);
 		this.scrollWidget = scrollWidget;
 		new ControlScrollBar(this);
