+++ src/main/java/binnie/craftgui/extratrees/dictionary/PagePlanksOverview.java
@@ -28,8 +28,8 @@ public class PagePlanksOverview extends PageAbstract<ItemStack> {
 	public void onValueChanged(final ItemStack species) {
 		this.deleteAllChildren();
 		final WindowAbstractDatabase database = Window.get(this);
-		new ControlText(this, new IArea(0.0f, 0.0f, this.size().x(), 24.0f), species.getDisplayName(), TextJustification.MiddleCenter);
-		new ControlText(this, new IArea(12.0f, 24.0f, this.size().x() - 24.0f, 24.0f), ExtraTrees.proxy.localise("gui.database.planks.use"), TextJustification.MiddleLeft);
+		new ControlText(this, new IArea(0, 0, this.size().x(), 24), species.getDisplayName(), TextJustification.MiddleCenter);
+		new ControlText(this, new IArea(12, 24, this.size().x() - 24, 24), ExtraTrees.proxy.localise("gui.database.planks.use"), TextJustification.MiddleLeft);
 		final IPlankType type = WoodManager.get(species);
 		int x = 12;
 		if (type != null) {
@@ -37,20 +37,20 @@ public class PagePlanksOverview extends PageAbstract<ItemStack> {
 			final ItemStack gate = WoodManager.getGate(type);
 			final ItemStack door = WoodManager.getDoor(type, DoorType.Standard);
 			if (fence != null) {
-				new ControlItemDisplay(this, x, 48.0f).setItemStack(fence);
+				new ControlItemDisplay(this, x, 48).setItemStack(fence);
 				x += 22;
 			}
 			if (gate != null) {
-				new ControlItemDisplay(this, x, 48.0f).setItemStack(gate);
+				new ControlItemDisplay(this, x, 48).setItemStack(gate);
 				x += 22;
 			}
 			if (door != null) {
-				new ControlItemDisplay(this, x, 48.0f).setItemStack(door);
+				new ControlItemDisplay(this, x, 48).setItemStack(door);
 				x += 22;
 			}
 		}
-		final ControlText controlDescription = new ControlText(this, new IArea(8.0f, 84.0f, this.getSize().x() - 16.0f, 0.0f), "", TextJustification.MiddleCenter);
-		final ControlText controlSignature = new ControlText(this, new IArea(8.0f, 84.0f, this.getSize().x() - 16.0f, 0.0f), "", TextJustification.BottomRight);
+		final ControlText controlDescription = new ControlText(this, new IArea(8, 84, this.getSize().x() - 16, 0), "", TextJustification.MiddleCenter);
+		final ControlText controlSignature = new ControlText(this, new IArea(8, 84, this.getSize().x() - 16, 0), "", TextJustification.BottomRight);
 		String desc = "";
 		if (type != null) {
 			desc = type.getDescription();
@@ -71,7 +71,7 @@ public class PagePlanksOverview extends PageAbstract<ItemStack> {
 		}
 		controlDescription.setValue(descBody + "§r");
 		controlSignature.setValue(descSig + "§r");
-		final float descHeight = CraftGUI.render.textHeight(controlDescription.getValue(), controlDescription.getSize().x());
-		controlSignature.setPosition(new IPoint(controlSignature.pos().x(), controlDescription.getPosition().y() + descHeight + 10.0f));
+		final int descHeight = CraftGUI.render.textHeight(controlDescription.getValue(), controlDescription.getSize().x());
+		controlSignature.setPosition(new IPoint(controlSignature.pos().x(), controlDescription.getPosition().y() + descHeight + 10));
 	}
 }
