+++ src/main/java/binnie/craftgui/extratrees/dictionary/WindowBrewery.java
@@ -18,7 +18,7 @@ import net.minecraftforge.fml.relauncher.Side;
 
 public class WindowBrewery extends Window {
 	public WindowBrewery(final EntityPlayer player, final IInventory inventory, final Side side) {
-		super(228.0f, 218.0f, player, inventory, side);
+		super(228, 218, player, inventory, side);
 	}
 
 	@Override
@@ -35,15 +35,15 @@ public class WindowBrewery extends Window {
 	public void initialiseClient() {
 		this.setTitle(Machine.getMachine(this.getInventory()).getPackage().getDisplayName());
 		new ControlSlotArray(this, 42, 32, 1, 3).create(BreweryMachine.SLOT_RECIPE_GRAINS);
-		new ControlSlot(this, 16.0f, 41.0f).assign(BreweryMachine.SLOT_RECIPE_INPUT);
-		new ControlSlot(this, 105.0f, 77.0f).assign(BreweryMachine.SLOT_YEAST);
+		new ControlSlot(this, 16, 41).assign(BreweryMachine.SLOT_RECIPE_INPUT);
+		new ControlSlot(this, 105, 77).assign(BreweryMachine.SLOT_YEAST);
 		new ControlLiquidTank(this, 76, 32).setTankID(BreweryMachine.TANK_INPUT);
 		new ControlLiquidTank(this, 162, 32).setTankID(BreweryMachine.TANK_OUTPUT);
 		new ControlEnergyBar(this, 196, 32, 16, 60, Position.Bottom);
-		new ControlBreweryProgress(this, 110.0f, 32.0f);
-		new ControlSlotArray(this, (int) (this.getSize().x() / 2.0f - 81.0f), 104, 9, 1).create(BreweryMachine.SLOTS_INVENTORY);
+		new ControlBreweryProgress(this, 110, 32);
+		new ControlSlotArray(this, this.getSize().x() / 2 - 81, 104, 9, 1).create(BreweryMachine.SLOTS_INVENTORY);
 		new ControlPlayerInventory(this);
-		new ControlErrorState(this, 133.0f, 79.0f);
+		new ControlErrorState(this, 133, 79);
 	}
 
 	public static Window create(final EntityPlayer player, final IInventory inventory, final Side side) {
