+++ src/main/java/binnie/craftgui/extratrees/kitchen/ControlDropDownMenu.java
@@ -8,8 +8,8 @@ import binnie.craftgui.window.Panel;
 public class ControlDropDownMenu extends Panel {
 	public boolean stayOpenOnChildClick;
 
-	public ControlDropDownMenu(final IWidget parent, final float x, final float y, final float width, final float height) {
-		super(parent, x, y, width, 2.0f, MinecraftGUI.PanelType.Gray);
+	public ControlDropDownMenu(final IWidget parent, final int x, final int y, final int width, final int height) {
+		super(parent, x, y, width, 2, MinecraftGUI.PanelType.Gray);
 		this.stayOpenOnChildClick = false;
 		this.addAttribute(Attribute.CanFocus);
 	}
