+++ src/main/java/binnie/craftgui/mod/database/ControlBranchBox.java
@@ -10,7 +10,7 @@ class ControlBranchBox extends ControlListBox<IClassification> {
 		return new ControlBranchBoxOption(this.getContent(), value, y);
 	}
 
-	public ControlBranchBox(final IWidget parent, final float x, final float y, final float width, final float height) {
-		super(parent, x, y, width, height, 12.0f);
+	public ControlBranchBox(final IWidget parent, final int x, final int y, final int width, final int height) {
+		super(parent, x, y, width, height, 12);
 	}
 }
